add deployment subcommand - change deploy to deployment update, include history refresh and promote
