+++ docs/Getting_Started.md
@@ -39,7 +39,7 @@ If this command throws no errors, try interacting with the running environment w
 ```bash
 hokusai local dev status
 hokusai local dev logs
-hokusai local dev shell
+hokusai local dev run bash
 ```
 
 To shut down the environment's running containers, run `hokusai local dev stop`
@@ -57,18 +57,18 @@ See [Configuration Options](./Configuration_Options.md) if you want to modify yo
 
 5) Build and push an image to ECR
 
-`hokusai local push` will build and push an image to ECR.  By default, it tags the image as the SHA1 of the HEAD of your current git branch (by calling `git rev-parse HEAD`).  You can override this behavior with the `--tag` option, although this is not recommended as creating builds matched to the SHA1 of Git tags gives you a clean view of your ECR project repository and deployment history.
+`hokusai registry push` will build and push an image to ECR.  By default, it tags the image as the SHA1 of the HEAD of your current git branch (by calling `git rev-parse HEAD`).  You can override this behavior with the `--tag` option, although this is not recommended as creating builds matched to the SHA1 of Git tags gives you a clean view of your ECR project repository and deployment history.
 
 The command will also tag the image as `latest`.  This image tag should not be referenced in any Kubernetes YAML configuration, but serves only as a pointer, which is referenced when creating a Kubernetes environment.
 
 The command aborts if any of the following conditions is met:
 - The working directory is not clean (you have uncommitted changes)
 - The working directory contains any files specified in your `.gitignore` file
-- The ECR project repository already contains the specified tag
+- The project registry already contains the specified tag
 
 The reason for these conditional checks is that when building, Docker will copy your _entire_ working directory into the container image, which can produce unexpected results when building images locally, destined for production environments!  Hokusai aborts if it detects the working directory is unclean, or any ignored files or directories are present, as it attempts to prevent any local configuration leaking into container images.
 
-Once an image is pushed, you can list the images and tags in the ECR project repository with: `hokusai remote images`.
+Once an image is pushed, you can list the images and tags in the project registry with: `hokusai registry images`.
 
 6) Create the Kubernetes staging environment and environment configuration
 
@@ -106,7 +106,7 @@ hokusai remote history --production
 
 10) Deploy changes to the Kubernetes staging environment
 
-Create a new commit and push it to the remote repo with `hokusai local push` as before.
+Create a new commit and push it to the remote repo with `hokusai registry push` as before.
 
 Deploy the new commit to staging with `hokusai remote deploy {TAG} --staging` where TAG is the tag of the commit you just made.
