+++ hokusai/cli/__init__.py
@@ -3,4 +3,5 @@ from hokusai.cli.local import local
 from hokusai.cli.local.dev import dev
 from hokusai.cli.registry import registry
 from hokusai.cli.remote import remote
+from hokusai.cli.remote.deployment import deployment
 from hokusai.cli.remote.env import env
