+++ hokusai/commands/__init__.py
@@ -1,17 +1,14 @@
 from hokusai.commands.build import build
 from hokusai.commands.check import check
 from hokusai.commands.configure import configure
-from hokusai.commands.deploy import deploy
+from hokusai.commands.deployment import update, history, refresh, promote
 from hokusai.commands.development import dev_start, dev_stop, dev_status, dev_logs, dev_run, dev_clean
 from hokusai.commands.gitdiff import gitdiff
 from hokusai.commands.gitlog import gitlog
 from hokusai.commands.env import create_env, delete_env, get_env, set_env, unset_env
-from hokusai.commands.history import history
 from hokusai.commands.images import images
 from hokusai.commands.logs import logs
-from hokusai.commands.promote import promote
 from hokusai.commands.push import push
-from hokusai.commands.refresh import refresh
 from hokusai.commands.run import run
 from hokusai.commands.setup import setup
 from hokusai.commands.remote_environment import environment_create, environment_update, environment_delete, environment_status
